+++ CHANGELOG.md
@@ -18,6 +18,17 @@ Categories, defined in [changemap.json](.github/clq/changemap.json):
   - `Fixed` for any bugfixes.
   - `Security` in case of vulnerabilities.
 
+## [1.1.0] - 2025-01-09
+
+### Added
+
+- publish single helm chart from `build/helm/charts/' if on the main branch
+- publish single docker image from `build/jib-image.tar' if on the main branch
+
+### Fixed
+
+- remove explicit call to helmPackage, it is part of the build sequence
+
 ## [1.0.0] - 2025-01-08
 
 ### Added